Rename CustomAuth to Auth.
