+++ config/auth_conf.xml.sample
@@ -1,6 +1,6 @@
 <?xml version="1.0"?>
-<customauth>
+<auth>
   <authenticator>
     <type>localdb</type>
   </authenticator>
-</customauth>
+</auth>
