+++ config/galaxy.ini.sample
@@ -836,12 +836,12 @@ use_interactive = True
 #openid_config_file = config/openid_conf.xml
 #openid_consumer_cache_path = database/openid_consumer_cache
 
-# XML config file that allows the use of custom authentication providers (e.g.
-# Active Directory) instead or in addition to local authentication (.sample is
-# used if default does not exist).
-#customauth_config_file = config/customauth_conf.xml
-# Print debugging info for customauth
-#customauth_debug = False
+# XML config file that allows the use of different authentication providers
+# (e.g. Active Directory) instead or in addition to local authentication
+# (.sample is used if default does not exist).
+#auth_config_file = config/auth_conf.xml
+# Print debugging info for auth
+#auth_debug = False
 
 # Optional list of email addresses of API users who can make calls on behalf of
 # other users.
