+++ lib/galaxy/auth/providers/localdb.yapsy-plugin
@@ -1,0 +1,8 @@
+[Core]
+Name = LocalDB
+Module = localdb
+
+[Documentation]
+Author = Andrew Robinson
+Version = 0.1
+Description = Authentication Provider that checks against the local database as per normal.
