+++ lib/galaxy/webapps/galaxy/controllers/mobile.py
@@ -1,7 +1,7 @@
 from galaxy import web
 from galaxy.web.base.controller import *
 
-import galaxy.customauth
+import galaxy.auth
 
 class Mobile( BaseUIController ):
 
@@ -59,7 +59,7 @@ class Mobile( BaseUIController ):
         # error = password_error = None
         # user = trans.sa_session.query( model.User ).filter_by( email = email ).first()
         # if not user:
-        #     autoreg = galaxy.customauth.check_auto_registration(trans, email, password, trans.app.config.customauth_config_file, trans.app.config.customauth_debug)
+        #     autoreg = galaxy.auth.check_auto_registration(trans, email, password, trans.app.config.auth_config_file, trans.app.config.auth_debug)
         #     if autoreg[0]:
         #         kwd = {}
         #         kwd['username'] = autoreg[1]
@@ -83,7 +83,7 @@ class Mobile( BaseUIController ):
         #     error = "This account has been marked deleted, contact your Galaxy administrator to restore the account."
         # elif user.external:
         #     error = "This account was created for use with an external authentication method, contact your local Galaxy administrator to activate it."
-        # elif not galaxy.customauth.check_password(user, password, trans.app.config.customauth_config_file, trans.app.config.customauth_debug):
+        # elif not galaxy.auth.check_password(user, password, trans.app.config.auth_config_file, trans.app.config.auth_debug):
         #     error = "Invalid password"
         # else:
         #     trans.handle_user_login( user )
